finished edit schedule for doctors, working in booking system

+++ app/src/main/java/com/kareem/yladoctor/Models/Modules/User/Individual.java
@@ -77,7 +77,7 @@ public class Individual extends User {
 		this.name = name;
 	}
 
-	public long getPoints () {
+	public int getPoints () {
 		return points;
 	}
